requirements.txt
torch
torchvision
Flask
Flask-Cors
Pillow
